requirements.txt
chromadb
flask
markdown
python-docx
